fix: Improve flakey E2E tests
