+++ e2e/tests/ui-driven/src/create-flow/helpers.ts
@@ -1,0 +1,39 @@
+import { Browser, Page, Request } from "@playwright/test";
+import { createAuthenticatedSession } from "../globalHelpers";
+
+export const getUserRequest = (req: Request) => {
+  const isHasuraRequest = req.url().includes("/graphql");
+  const isGetUserRequest = Boolean(
+    isHasuraRequest && req.postData()?.toString().includes("GetUserById"),
+  );
+  return isGetUserRequest;
+};
+
+export async function getAdminPage({
+  browser,
+  userId,
+}: {
+  browser: Browser;
+  userId: number;
+}): Promise<Page> {
+  const page = await createAuthenticatedSession({ browser, userId });
+  await page.goto("/");
+  await page.waitForResponse((response) => {
+    return response.url().includes("/graphql");
+  });
+  return page;
+}
+
+export async function getTeamPage({
+  browser,
+  userId,
+  teamName,
+}: {
+  browser: Browser;
+  userId: number;
+  teamName: string;
+}): Promise<Page> {
+  const page = await getAdminPage({ browser, userId });
+  await page.locator("h2", { hasText: teamName }).click();
+  return page;
+}
